Fixed validation of password confirm Used ReactiveForm mode and FormGroup validator. Still working on template mode and directive validation on multiple controls

+++ src/app/+user/user.component.ts
@@ -1,6 +1,6 @@
-import { Inject, Component, OnInit, Input } from '@angular/core';
+import { Inject, Component, OnInit, Input, ViewContainerRef } from '@angular/core';
 import { Router, ActivatedRoute, Params } from '@angular/router';
-import { MdDialog, MdDialogRef } from '@angular/material/dialog';
+import { MdDialog, MdDialogRef, MdDialogConfig } from '@angular/material/dialog';
 import { ResetPasswordDialogComponent } from '../dialogs/reset-password-dialog/reset-password-dialog.component';
 import { Location } from '@angular/common';
 import { IUser } from '../shared/IUser';
@@ -8,6 +8,7 @@ import { User } from '../shared/User';
 
 import { UserServiceBase, USER_SERVICE_TOKEN } from '../IServices/IUserService';
 import { JWTServiceBase, JWT_SERVICE_TOKEN } from '../IServices/IJWTService';
+import { NotificationsService } from '../Services/notifications.service';
 
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/switchMap';
@@ -23,31 +24,36 @@ export class UserComponent implements OnInit {
 
     @Input()
     userId: string;
+    dialogConfig: MdDialogConfig = new MdDialogConfig();
 
     user: IUser = new User();
     availableRoles: string[] = [];
     availableApps: string[] = [];
 
-    error: string;
-
-    debugMode:boolean = environment.debugMode;
+    debugMode: boolean = environment.debugMode;
 
     constructor(
         private activeRoute: ActivatedRoute,
         private location: Location,
         @Inject(USER_SERVICE_TOKEN) private userService: UserServiceBase,
         @Inject(JWT_SERVICE_TOKEN) private jwtService: JWTServiceBase,
-        private dialog:MdDialog,
-        private router: Router
-    ) { }
+        private dialog: MdDialog,
+        private toaster: NotificationsService,
+        private router: Router,
+        private vcr:ViewContainerRef
+    ) {
+        this.dialogConfig.role = 'dialog';
+        this.dialogConfig.viewContainerRef = this.vcr;
+
+    }
 
     ngOnInit() {
+        this.toaster.ViewContainerRef = this.vcr;
 
         if (!this.jwtService.is_authenticated) // jwt token not exists, navigate to login view 
             this.router.navigate(['/login']);
 
         this.userId = this.activeRoute.snapshot.params['name'];
-        this.error = '';
 
         let self = this;
 
@@ -55,25 +61,25 @@ export class UserComponent implements OnInit {
             .then(user => {
                 self.user = user;
                 console.log(`User ${this.userId} loaded`);
-            }, error => self.error = error._body);
+            }, error => this.toaster.showToastError(error));
 
         this.userService.getAvailableApplications()
             .then(value => self.availableApps = value,
-            error => this.error += `\n${error._body}`);
+            error => this.toaster.showToastError(error));
 
         this.userService.getAvailableRoles()
             .then(value => self.availableRoles = value,
-            error => this.error += `\n${error._body}`);
+            error => this.toaster.showToastError(error));
     }
 
     enableAllApplications() {
-        this.user.applications = [];        
-        this.availableApps.forEach( app => this.user.applications.push(app));
+        this.user.applications = [];
+        this.availableApps.forEach(app => this.user.applications.push(app));
     }
 
     enableAllRoles() {
         this.user.roles = [];
-        this.availableRoles.forEach(role => this.user.roles.push(role) );
+        this.availableRoles.forEach(role => this.user.roles.push(role));
     }
 
     toggleApplication(app: string) {
@@ -85,7 +91,7 @@ export class UserComponent implements OnInit {
     }
 
     isInRole(role: string): boolean {
-        return this.user.roles.indexOf(role) >  -1;
+        return this.user.roles.indexOf(role) > -1;
     }
 
     isInApplication(app: string): boolean {
@@ -102,10 +108,12 @@ export class UserComponent implements OnInit {
 
     }
 
-    delete(){
-        this.userService.delete(this.user._id)
-            .then(() => this.router.navigate(['/users']))
-            .catch(error => this.error = error._body || error);                
+    delete() {
+        if (window.confirm('Are you sure to delete current user?')){
+            this.userService.delete(this.user._id)
+                .then(() => this.router.navigate(['/users']))
+                .catch(error => this.toaster.showToastError(error));
+        }
     }
 
     update() {
@@ -113,20 +121,19 @@ export class UserComponent implements OnInit {
         this.userService.updateUser(this.userId, this.user)
             .then((val: IUser) => {
                 console.debug('user update succeed');
-                this.router.navigate( ['/']);
+                this.router.navigate(['/']);
             })
-            .catch(err => console.error(err));
+            .catch(err => this.toaster.showToastError(err));
 
     }
 
-    changePassword(){
-        let resetPassword:MdDialogRef<ResetPasswordDialogComponent> = this.dialog.open(ResetPasswordDialogComponent);
-        resetPassword.afterClosed().subscribe((result) => 
-        {
-            if (result != null){
+    changePassword() {
+        let resetPassword: MdDialogRef<ResetPasswordDialogComponent> = this.dialog.open(ResetPasswordDialogComponent, this.dialogConfig);
+        resetPassword.afterClosed().subscribe((result) => {
+            if (result != null) {
                 this.userService.resetPassword(this.userId, result.password, result.password_confirm)
-                    .then(()=> this.error = undefined)
-                    .catch((error) => this.error = error.message || error._body || error);
+                    .then(() => this.toaster.showToastInfo('Password changed'))
+                    .catch((error) => this.toaster.showToastError(error));
             }
             resetPassword = null;
         });
